requirements.txt
django==6.0.3
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
django-filter==25.2
drf-spectacular==0.29.0
python-decouple==3.8
Pillow==12.1.1
